SNN + GA vectorised rewrite
I have implemented a vectorised SNN, massively simplified compared to the previous version. This is not quite complete (in that it is a forward pass tied to the simulation rather than a free-flowing stream). The GA has also been rewritten, again with a focus on simplicity. There some necessary additions here, particularly in waiting for the MLX fix for pickling. I will also implement novelty search soon.